move stateWrapper to own file + only expose state object through wrapper
